# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Apache Airavata HTTP API
  version: "1.0.0"
  description: |
    The REST API served by `airavata-server` (see `internal/server/server.go`).

    All request and response bodies are JSON (`Content-Type: application/json`).

    ## Authentication

    Authentication is a bearer token validated by CILogon introspection, carried as
    `Authorization: Bearer <token>`. A request without a token is anonymous rather
    than rejected at the door: each operation decides for itself whether an anonymous
    caller is acceptable. A token that is present but invalid is always `401`, with
    `WWW-Authenticate: Bearer error="invalid_token"`. If the identity provider cannot
    be reached the call fails `502`, because that is not the caller's fault.

    ## Authorization

    Two schemes are in play, depending on the resource:

    - **Platform roles** (`ADMIN`, `SUPER_ADMIN`) guard the shared catalogue —
      Slurm clusters, partitions, application templates and batch deployments.
      Reads of the catalogue are open to anyone, including anonymous callers;
      writes require a role.
    - **Ownership and sharing** guards the per-user resources — SSH keys, groups,
      cluster configs, SCP data storages, data products and processes. Ownership
      comes from the token and is never carried in a request body, so nothing can be
      created on someone else's behalf or handed over by editing it. None of these
      are readable anonymously.

...

...

...

...
    state.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/ssh-keys:
    get:
      tags: [SSH Keys]
      summary: List the caller's SSH keys
      description: Returns only the caller's own keys. Admins are not exempt.
      operationId: listSSHKeys
      responses:
        "200":
          description: The caller's keys.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SSHKeyResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [SSH Keys]
      summary: Register an SSH key
      description: |
        Self-service for any authenticated caller with a `users` row. `privateKey` is
        required on create; `privateKey` and `passphrase` are write-only and are never
        returned by any read endpoint.
      operationId: createSSHKey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SSHKeyRequest" }
      responses:
        "201":
          description: The registered key, without its private material.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SSHKeyResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/ssh-keys/{sshKeyId}:
    parameters:
      - $ref: "#/components/parameters/SSHKeyId"
    get:
      tags: [SSH Keys]
      summary: Read an SSH key
      description: Owner only.
      operationId: getSSHKey
      responses:
        "200":
          description: The key, without its private material.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SSHKeyResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [SSH Keys]
      summary: Update an SSH key
      description: |
        Owner only. An update that omits `privateKey` keeps the stored secret.
      operationId: updateSSHKey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SSHKeyRequest" }
      responses:
        "200":
          description: The updated key.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SSHKeyResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [SSH Keys]
      summary: Delete an SSH key
      description: |
        Owner only. A key still referenced by a cluster config or a data storage
        cannot be deleted: `409 Conflict`.
      operationId: deleteSSHKey
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ------------------------------------------------------- Slurm clusters
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A CILogon access token, validated by RFC 7662 introspection. The bootstrap
        root token is also accepted; see INSTALL.md.
  parameters:
...
    SSHKeyId:
      name: sshKeyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NoContent:
      description: Deleted. No body.
    BadRequest:
      description: |
        Malformed JSON, or a body that failed its constraints. Validation failures
        carry `fieldErrors`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            validation:
              summary: Validation failed
              value:
                status: 400
                error: Bad Request
                message: Validation failed
                fieldErrors:
                  - field: clusterName
                    message: Cluster name cannot be blank
    Unauthorized:
      description: |
        No token where one is required, or a token that is present but unusable. An
        invalid token also carries `WWW-Authenticate: Bearer error="invalid_token"`.
      headers:
        WWW-Authenticate:
          description: Present when the token was rejected.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated, but lacking the authority for this call.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such record — or, for a group, a record the caller has no standing to know
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            status: 404
            error: Not Found
            message: "Slurm cluster not found: c1d2e3f4-5678-4abc-9def-0123456789ab"
    Conflict:
      description: |
        The request collides with existing state — deleting an SSH key still in use,
        or a template that still has deployments.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: |
        The identity provider could not be reached to validate the token. Not the
        caller's fault, so it is not reported as an auth failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: |
        An unexpected server failure. The message is always `"Internal server error"`;
        the detail is logged rather than returned, since it may name internal state.

        A `502` is also possible on any authenticated call if the identity provider is
        unreachable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    FieldError:
      type: object
      description: One failed validation constraint.
      properties:
        field:
          type: string
          description: |
            The JSON path of the offending field, including indexes and nesting.
          examples:
            ["inputs[0].inputType", "defaultBatchJobConfig.wallTimeMinutes"]
        message: { type: string }
      required: [field, message]
    Error:
      type: object
      description: The envelope every failure response uses.
      properties:
        status:
          type: integer
          description: The HTTP status, repeated in the body.
        error:
          type: string
          description: The standard reason phrase for `status`.
        message:
          type: string
          description: Human-readable detail.
        fieldErrors:
          type: array
          description: |
            Present only on request-body validation failures, listing every
            constraint that failed so a client can report all of them at once.
          items: { $ref: "#/components/schemas/FieldError" }
      required: [status, error, message]

    # ----------------------------------------------------------- Enums
...
    SSHKeyRequest:
      type: object
      description: |
        The create/update payload for a key. `privateKey` is required on create; an
        update that omits it keeps the stored secret. There is no owner field.
      properties:
        sshKeyName: { type: string }
        publicKey: { type: string }
        privateKey:
          type: [string, "null"]
          description: Write-only. Never returned by any read endpoint.
        passphrase:
          type: [string, "null"]
          description: Write-only. Never returned by any read endpoint.
      required: [sshKeyName, publicKey]
    SSHKeyResponse:
      type: object
      description: |
        The read model for a key. It has no private key or passphrase field at all —
        the secrets cannot leak through this endpoint because there is nowhere for
        them to go.
      properties:
        sshKeyId: { type: string, format: uuid }
        sshKeyName: { type: string }
        publicKey: { type: string }
        ownerId: { type: string }
      required: [sshKeyId, sshKeyName, publicKey, ownerId]

    # -------------------------------------------------------- Compute
...